qmui2.x适配中

+++ ajs/src/main/java/com/jsongo/ajs/interaction/Topbar.kt
@@ -83,7 +83,7 @@ object Topbar {
             if (textSizeStr != null) {
                 val toDoublet = textSizeStr.toDoubleOrNull()
                 toDoublet?.let {
-                    topbarTitle.textSize = it.toFloat()
+                    topbarTitle.textSize = it.toInt()
                 }
             }
             callback.success()
